Remove a bad cast tripping up -Wsign-conversion
Prolly needs a test routine.  Also prolly shoul return an int64_t not
just a long.

+++ libparse/binio.c
@@ -42,7 +42,7 @@ get_lsb_long(
   retval  = *((*bufpp)++);
   retval |= *((*bufpp)++) << 8;
   retval |= *((*bufpp)++) << 16;
-  retval |= (unsigned long)*((*bufpp)++) << 24;
+  retval |= *((*bufpp)++) << 24;
 
   return retval;
 }
